requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.1
blis==0.7.9
cachetools==5.3.0
catalogue==1.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
cymem==2.0.7
Django==4.0.8
django-cors-headers==3.13.0
django-crontab==0.7.1
django-elasticsearch-dsl==7.3
django-phonenumber-field==7.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
en-core-web-sm==2.3.1
errant==2.3.3
filelock==3.9.0
fsspec==2022.11.0
fuzzywuzzy==0.18.0
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
gramformer==1.0
httplib2==0.21.0
huggingface-hub==0.11.1
idna==3.4
Levenshtein==0.20.9
murmurhash==1.0.9
numpy==1.24.1
packaging==22.0
phonenumberslite==8.13.3
Pillow==9.3.0
plac==1.1.3
preshed==3.0.8
protobuf==4.21.12
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytz==2022.7
PyYAML==6.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.1
rsa==4.9
sentencepiece==0.1.97
six==1.16.0
soupsieve==2.4
spacy==2.3.9
sqlparse==0.4.3
srsly==1.0.6
thinc==7.4.6
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
wasabi==0.10.1
